+++ desktop-player/src/preload.js
@@ -1,0 +1,7 @@
+"use strict";
+
+const { contextBridge, ipcRenderer } = require("electron");
+
+contextBridge.exposeInMainWorld("ajrmPlayer", {
+  appVersion: () => ipcRenderer.invoke("app-version"),
+});